Move all authentication classes to the authentication subdirectory
