+++ Questrade/DependencyLoader.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
+using ParityService.Questrade.Authentication;
 using System;
 
 namespace ParityService.Questrade
